Changes: Added the ability to decrypt using the GUI
Bugs;
After you generate a key, it takes a while before it shows up in the encrypt key choice menu
While changing the size of the GUI, the button sizes get messed up

Need to Implement:
Ability to generate a key within the encrypt menu
Manage Key Pairs table to work with data from the database. Also, the ability to import a key/key pair from there.
Password based encryption as another option
Exporting of public keys to a file so it can easily be sent to another user who will then import it and use it for encryption
UI beautification